pyproject.toml
[project]
name = "back-project"
version = "0.1.0"
description = ""
authors = [
    { name = "alexkostromin", email = "alexkostromin9797@gmail.com" }
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.30.0,<1.0.0)",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "uuid7 (>=0.1.0,<1.0.0)",
    "structlog (>=24.0.0,<26.0.0)",
    "aiofiles (>=24.0.0,<25.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
